refactor(config): refactor mask_rcnn/ to use pydantic config
