+++ src/autovisionai/models/mask_rcnn/mask_rcnn_model.py
@@ -2,7 +2,7 @@ import torchvision
 from torchvision.models.detection.faster_rcnn import FastRCNNPredictor
 from torchvision.models.detection.mask_rcnn import MaskRCNN_ResNet50_FPN_Weights, MaskRCNNPredictor
 
-from autovisionai.configs.config import CONFIG
+from autovisionai.configs import CONFIG
 
 
 def create_model(n_classes: int, pretrained: bool = True) -> MaskRCNNPredictor:
@@ -24,7 +24,7 @@ def create_model(n_classes: int, pretrained: bool = True) -> MaskRCNNPredictor:
 
     # Getting the number of input features for the mask classifier
     in_features_mask = model.roi_heads.mask_predictor.conv5_mask.in_channels
-    hidden_layer = CONFIG["mask_rcnn"]["model"]["hidden_size"].get()
+    hidden_layer = CONFIG.models.mask_rcnn.hidden_size
 
     model.roi_heads.mask_predictor = MaskRCNNPredictor(
         in_channels=in_features_mask, dim_reduced=hidden_layer, num_classes=n_classes
